requirements.txt
python-dotenv
pillow